Fixes issue where deleting and creating product properties at the same time caused an error on the product page stating the product could not be found.
When deleting a product property it was deleted from the database and the row was hidden (not removed) on the frontend. On submitting the form, the now deleted property id was still present causing the error when trying to load the property with that id from the database.

Fixes #4467

Conflicts:
	backend/spec/features/admin/products/properties_spec.rb