Delete hidden line item id input when removing tr
This ensures we don't send the id of the line item that got removed
through ajax. If the input is not removed from the DOM the backend tries
to load the line item and fails to find it.

This won't be probably needed in Spree 2.0 as then the
admin/orders/_form template does not seem to add hidden inputs with the
ids of the line items.