Do not render HTML when asking for JS
As it is, Spree always renders the HTML form which gets rendered within
a flash error in OFN. That is because returning HTML for a JS ajax call
raises a 'parsererror' thus, rendering it in the view.

By doing this change, we force Spree to render the destroy.js.erb view.